rename the jaxrs + jaxws interop test
