+++ dev/io.openliberty.ws.jaxrs.2.1.jaxws.2.2_fat/.classpath
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
 	<classpathentry kind="src" path="fat/src"/>
-	<classpathentry kind="src" path="test-applications/jaxrs-prototype/src"/>
+	<classpathentry kind="src" path="test-applications/jaxrs-interop/src"/>
 	<classpathentry kind="src" path="test-applications/jaxws-peopleservice/src"/>
 	<classpathentry kind="con" path="aQute.bnd.classpath.container"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
