+++ dev/io.openliberty.ws.jaxrs.2.1.jaxws.2.2_fat/bnd.bnd
@@ -1,10 +1,10 @@
 #*******************************************************************************
-# Copyright (c) 2022 IBM Corporation and others.
+# Copyright (c) 2022, 2023 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
 # http://www.eclipse.org/legal/epl-2.0/
-# 
+#
 # SPDX-License-Identifier: EPL-2.0
 #
 # Contributors:
@@ -15,7 +15,7 @@ bVersion=1.0
 
 src: \
   fat/src,\
-  test-applications/jaxrs-prototype/src,\
+  test-applications/jaxrs-interop/src,\
   test-applications/jaxws-peopleservice/src
 
 fat.project: true
@@ -26,4 +26,4 @@ fat.project: true
   com.ibm.websphere.javaee.jaxws.2.2;version=latest,\
   com.ibm.websphere.javaee.jaxrs.2.0,\
   fattest.simplicity
-	
+
