+++ dev/io.openliberty.ws.jaxrs.2.1.jaxws.2.2_fat/fat/src/com/ibm/ws/fat/WebServicesInteroperabilityTest.java
@@ -1,10 +1,10 @@
 /*******************************************************************************
- * Copyright (c) 2022 IBM Corporation and others.
+ * Copyright (c) 2022, 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
@@ -19,13 +19,13 @@ import org.junit.BeforeClass;
 import org.junit.runner.RunWith;
 
 import com.ibm.websphere.simplicity.ShrinkHelper;
-import com.ibm.ws.jaxrs.fat.prototype.InteropStartClientTestServlet;
 
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
 import componenttest.custom.junit.runner.FATRunner;
 import componenttest.topology.impl.LibertyServer;
 import componenttest.topology.utils.FATServletClient;
+import io.openliberty.interop.InteropStartClientTestServlet;
 
 /*
  * The purpose of this test is to check interoperability between JAX-RS and JAX-WS
@@ -36,7 +36,7 @@ import componenttest.topology.utils.FATServletClient;
 @RunWith(FATRunner.class)
 public class WebServicesInteroperabilityTest extends FATServletClient {
 
-    private static final String appName = "prototype";
+    private static final String appName = "interop";
 
     @Server("WebServicesInteroperabilityServer")
     @TestServlet(servlet = InteropStartClientTestServlet.class, contextRoot = appName)
@@ -45,7 +45,7 @@ public class WebServicesInteroperabilityTest extends FATServletClient {
     @BeforeClass
     public static void setup() throws Exception {
         // Build an application and export it to the dropins directory
-        ShrinkHelper.defaultDropinApp(server, appName, "com.ibm.ws.jaxrs.fat.prototype",
+        ShrinkHelper.defaultDropinApp(server, appName, "io.openliberty.interop",
                                       "com.ibm.ws.jaxws.test.wsr.server.stub",
                                       "com.ibm.ws.jaxws.fat.util");
