+++ dev/io.openliberty.ws.jaxrs.2.1.jaxws.2.2_fat/test-applications/jaxrs-interop/src/io/openliberty/interop/InteropStartJaxrsResource.java
@@ -1,22 +1,22 @@
 /*******************************************************************************
- * Copyright (c) 2022 IBM Corporation and others.
+ * Copyright (c) 2022, 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.jaxrs.fat.prototype;
+package io.openliberty.interop;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 
 @Path("/ep2")
-public class PrototypeResource {
+public class InteropStartJaxrsResource {
 
     @GET
     @Path("jaxwsEP2")
